Convert Resource class model to TypeScript
