+++ package.json
@@ -39,7 +39,7 @@
     "build-storybook": "yarn run install-storybook && build-storybook -s .storybook/public,assets",
     "install-storybook": "./scripts/storybook-install",
     "remove-storybook": "./scripts/storybook-install -d",
-    "remove-storybook-deps": "./scripts/storybook-install -r"    
+    "remove-storybook-deps": "./scripts/storybook-install -r"
   },
   "dependencies": {
     "@aws-sdk/client-ec2": "^3.1.0",
@@ -65,6 +65,7 @@
     "dayjs": "^1.8.16",
     "diff2html": "^2.11.2",
     "dompurify": "^2.0.12",
+    "eslint-plugin-n": "^15.2.1",
     "event-target-shim": "^5.0.1",
     "express": "^4.17.1",
     "file-saver": "^2.0.2",
@@ -151,8 +152,8 @@
     "eslint-import-resolver-node": "0.3.4",
     "eslint-module-utils": "2.6.1",
     "eslint-plugin-cypress": "^2.11.3",
-    "eslint-plugin-jest": "^23.6.0",
     "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-jest": "^23.6.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-vue": "7.14.0",
     "frontmatter-markdown-loader": "^3.6.3",
@@ -164,12 +165,12 @@
     "serve-static": "^1.14.1",
     "start-server-and-test": "^1.13.1",
     "style-loader": "^1.1.2",
+    "ts-jest": "^27.1.3",
     "typescript": "~4.1.5",
     "vue": "2.6.14",
     "vue-template-compiler": "2.6.14",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-virtual-modules": "^0.4.3",
-    "ts-jest": "^27.1.3",
     "yarn": "^1.22.11"
   },
   "nyc": {
