+++ pkg/epinio/pages/index.vue
@@ -8,10 +8,11 @@ import { EPINIO_MGMT_STORE, EPINIO_TYPES } from '../types';
 import Resource from '@shell/plugins/dashboard-store/resource-class';
 import AsyncButton from '@shell/components/AsyncButton.vue';
 import { _MERGE } from '@shell/plugins/dashboard-store/actions';
+import { STATES_ENUM, STATE_TYPE } from '@shell/plugins/dashboard-store/types/rancher-api-types';
 
 interface Cluster extends Resource{
   id: string,
-  state: string,
+  state: STATES_ENUM
 }
 
 interface Data {
@@ -46,7 +47,7 @@ export default Vue.extend<Data, any, any, any>({
     },
 
     canRediscover() {
-      return !this.clusters.find((c: Cluster) => c.state === 'updating');
+      return !this.clusters.find((c: Cluster) => c.state === STATES_ENUM.UPDATING);
     },
 
     clusters() {
