+++ shell/plugins/dashboard-store/types/router-types.ts
